Time shifter complete
After the pcap is filtered and exported, the timestamp difference between two
nearest packets of the pcap can have a large gap.

This tool helps to fill this gap by shifting the packet backwards. This is done
by re-writing the timestamp of each packet in 20mS difference starting from 0.
This will be helpful to use pcaps for pumping RTP from SIPp